add ts to health module
