--- src/modules/health/route.js
@@ -1,9 +1,0 @@
-const express = require('express');
-const constroller = require('./controller');
-
-const router = express.Router();
-
-router.get('/health', constroller.health);
-router.get('/is_healthy', constroller.is_healthy);
-
-module.exports = router;
